requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.7.2
chromadb==0.5.3
sentence-transformers==3.0.1
pdfplumber==0.11.1
python-dotenv==1.0.1
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0